Upgrade mariadb Helm chart to 8.0.0
The names of the values changed in the new Helm chart, see [1]. Thus the
Helm chart ships breaking changes and therefore bump the major version
of this Chart.yaml (not sure if you use semver?)

Furthermore, there is a bug [2] where not setting the root password
results in the mariadb entering a CrashLoopBackOff when re-installing
it. Thus set it (as done in the mysql values already).

[1] https://github.com/bitnami/charts/commit/1206942b69c02a140a60e4e5ab59bf6d53d5b7e6
[2] https://github.com/bitnami/charts/issues/3884