Introduce NPM dependencies for development
We use `readme-generator-for-helm` and `markdownlint-cli` which are both
NPM packages. So they should be available for developers in a more usual
way: `package.json` and `package-lock.json`.

As the currently used `readme-generator-for-helm` from npmjs.com is not
officially maintained by Bitnami (see issue link below), the change also
includes switching to the official GitHub repository.

https://github.com/bitnami-labs/readme-generator-for-helm/issues/36

Signed-off-by: justusbunsi <sk.bunsenbrenner@gmail.com>