Integrate readme integrity check into CI
With the recent switch to auto-generate the parameters inside README.md,
contributors should integrate an updated README.md into the pull request.

A Drone build now ensures that the README.md is properly updated by
running the script a contributor should run and checks for uncommitted
changes for README.md. If this produces a change, the build fails.

Signed-off-by: justusbunsi <sk.bunsenbrenner@gmail.com>