Improve the reading flow of upgrading notes
Every version section starts with a disclaimer right now. This is
duplicated and might hide important text due to its existence. A
centralized intruduction at top of the upgrading section tells the
reader what to expect.

I've noticed that on ArtifactHub the emotes are not rendered correctly.
So I replaced them with the actual ones and reduced their usage. That
way it better highlights those parts the reader must not miss.

Signed-off-by: justusbunsi <sk.bunsenbrenner@gmail.com>