Pre-generate LFS secret token during init phase
When enabling the LFS server feature, Gitea tries to generate a new LFS
token and persist it to the app.ini file, if no token is provided. In
the context of this Helm Chart, the file permissions do not allow the
"migrate" command to change this file which leads to write errors.

It doesn't harm to generate such token during init phase like it's done
with the INTERNAL_TOKEN and other non-changable values. Doing so, Gitea
won't try to persist a new token value and the Helm Chart succeeds its
upscale.

Signed-off-by: justusbunsi <sk.bunsenbrenner@gmail.com>